Fixed some files, and created basic search
