+++ Login-/Admin Panel/CRUD/images/index.php
@@ -1,5 +1,5 @@
 <?php
-require_once '../../dbconn.php';
+require_once '../dbConn/dbconn.php';
 try {
     $pdo = new PDO($attr, $user, $pass, $opts);
 } catch (PDOException $e) {
