--- Login-/Admin Panel/crud.php
@@ -1,3 +1,0 @@
-<?php 
-phpversion()
-?>
