--- Login-/Admin Panel/dbconn.php
@@ -1,14 +1,0 @@
-<?php
-    $host = 'localhost';
-    $data = 'job_orbitz';
-    $user = 'root';
-    $pass = '';
-    $chrs = 'utf8mb4';
-    $attr = "mysql:host=$host;dbname=$data;charset=$chrs";
-    $opts = 
-    [
-        PDO:: ATTR_ERRMODE              => PDO::ERRMODE_EXCEPTION,
-        PDO:: ATTR_DEFAULT_FETCH_MODE   => PDO::FETCH_ASSOC,
-        PDO:: ATTR_EMULATE_PREPARES      => false,
-    ];  
-?>
