+++ Login-/Admin Panel/logout.php
@@ -4,6 +4,6 @@
     session_unset();
     session_destroy();
 
-    header("location: index.php");
+    header("location: login.php");
     exit;
 ?>
