--- Login-/Website/mobile.css
@@ -1,70 +1,0 @@
-@media screen and (max-width: 650px) {
-    html {
-        font-size: 70%;
-    }
-
-    .btn {
-        padding: .5rem 1rem;
-    }
-
-    .topNav {
-        padding: 0 1rem;
-    }
-
-    .logo>img {
-        width: 150px;
-    }
-
-    .hero {
-        padding: 0 1rem;
-    }
-
-    .heroLeft {
-        gap: 2rem;
-    }
-
-    .heroLeft h1 {
-        font-size: 17.3vw;
-        line-height: 15vw;
-    }
-
-    .searchForm {
-        width: 100%;
-        padding: .2rem;
-    }
-
-    .searchForm label {
-        padding: 0 0 0 .5rem;
-    }
-
-    .searchForm select {
-        display: none;
-    }
-
-    .job {
-        flex-direction: column;
-        gap: 2rem;
-    }
-
-    .members {
-        height: 450px;
-    }
-
-    .member {
-        flex-direction: column;
-        text-align: center;
-    }
-
-    .footer-top .wrapper {
-        flex-direction: column;
-    }
-
-    .f-t-about {
-        width: 100%;
-    }
-
-    .doubleMenu {
-        justify-content: flex-start;
-        gap: 6rem;
-    }
-}
