--- Login-/Website/styles.css
@@ -1,473 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@100;200;300;400;500;600;700;800;900&display=swap');
-
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    text-decoration: none;
-    list-style-type: none;
-    font-family: 'Poppins', sans-serif;
-}
-
-html {
-    scroll-behavior: smooth;
-    scrollbar-width: none;
-}
-
-::selection {
-    background-color: none;
-    color: none;
-}
-
-.btn {
-    all: unset;
-    padding: .5rem 2rem;
-    border-radius: 10px;
-    background-color: #566bfc;
-    color: white;
-    cursor: pointer;
-}
-
-.wrapper {
-    max-width: 1368px;
-    margin: auto;
-    height: 100%;
-}
-
-header {
-    height: 100vh;
-    min-height: 550px;
-    max-height: 700px;
-    background-color: #566bfc;
-}
-
-.topNav {
-    height: 100px;
-    display: flex;
-    justify-content: space-between;
-    padding: 0 2rem;
-    align-items: center;
-}
-
-.logo>img {
-    width: 200px;
-}
-
-.hamburger {
-    height: 30px;
-    width: 40px;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-around;
-    align-items: flex-end;
-    cursor: pointer;
-    display: none;
-}
-
-.line {
-    height: 5px;
-    width: 100%;
-    background-color: white;
-    border-radius: 10px;
-}
-
-.line:nth-child(2) {
-    width: 60%;
-}
-
-.mainMenu {
-    flex: 1;
-    display: flex;
-    align-items: flex-start;
-    justify-content: space-between;
-    padding-bottom: .5rem;
-}
-
-.navLinks {
-    padding: 0 10rem;
-    display: flex;
-    justify-content: center;
-    gap: 3rem;
-    font-weight: 500;
-}
-
-.navLinks a {
-    color: white;
-}
-
-.navBtn {
-    background-color: white;
-    font-weight: 700;
-    padding: .5rem 3rem;
-    color: #566bfc;
-}
-
-/* hero */
-.hero {
-    display: flex;
-    justify-content: center;
-    padding: 0 2rem;
-    overflow-y: hidden;
-    height: calc(100% - 100px);
-}
-
-.heroLeft {
-    flex: 2;
-    z-index: 1;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-}
-
-.heroLeft h1 {
-    font-size: 12vw;
-    line-height: 10vw;
-    font-weight: 900;
-    color: white;
-}
-
-.searchForm {
-    background-color: white;
-    border-radius: 10px;
-    display: flex;
-    justify-content: center;
-    align-content: center;
-    height: 3.5rem;
-    padding: .5rem.5rem .5rem 1rem;
-    gap: 1rem;
-    margin-top: 1rem;
-}
-
-.searchForm label {
-    display: flex;
-    align-items: center;
-}
-
-.searchForm label img {
-    width: 15px;
-}
-
-.searchForm input {
-    all: unset;
-    flex: 1;
-}
-
-.searchForm select {
-    width: 120px;
-    background-color: transparent;
-    border: none;
-    border-left: 1px solid grey;
-    border-right: 1px solid grey;
-    padding: 0 1rem;
-}
-
-.heroRight {
-    flex: 1;
-    position: relative;
-    height: 100%;
-}
-
-.heroRight img {
-    width: 120%;
-    position: absolute;
-    left: -70px;
-}
-
-
-/* features */
-.features .wrapper {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    min-height: 400px;
-    flex-wrap: wrap;
-    gap: 2rem;
-    background-color: white;
-    padding: 2rem;
-}
-
-.feature {
-    width: 300px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.feature img {
-    width: 50px;
-    margin-bottom: 2rem;
-}
-
-.feature h2 {
-    font-weight: 400;
-    text-align: center;
-}
-
-.feature p {
-    text-align: center;
-    color: grey;
-}
-
-/* latest jobs */
-.container .wrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: 4rem 0;
-}
-
-.latest-jobs {
-    background: linear-gradient(to top, white, #566bfc);
-    color: white;
-}
-
-.container-heading {
-    font-size: 3rem;
-    text-align: center;
-    font-weight: 500;
-}
-
-.container-sub-heading {
-    font-size: 1.5rem;
-    text-align: center;
-    font-weight: 400;
-    margin-bottom: 4rem;
-}
-
-.jobs-list {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    align-items: center;
-    width: 100%;
-    max-width: 900px;
-    color: black;
-}
-
-.job {
-    background-color: white;
-    width: 90%;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 1rem;
-    border-radius: 10px;
-    min-height: 160px;
-    gap: 1rem;
-    box-shadow: 0 0 5px #566cfc38;
-}
-
-.job-details {
-    display: flex;
-    flex-direction: column;
-    flex: 1;
-    gap: 2rem;
-}
-
-.job-details h2 {
-    font-size: 2rem;
-    font-weight: 500;
-}
-
-.l-d-v {
-    display: flex;
-    gap: 2rem;
-}
-
-.l-d-v p {
-    display: flex;
-    align-items: center;
-    gap: .6rem;
-    font-size: 1.2rem;
-    color: grey;
-}
-
-.l-d-v p img {
-    width: 1.2rem;
-}
-
-.jobsBtn {
-    margin-top: 3rem;
-}
-
-
-/* Team */
-.team .wrapper {
-    background-color: white;
-}
-
-.team h2 {
-    color: #566bfc;
-}
-
-.members {
-    width: 100%;
-    max-width: 800px;
-    position: relative;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    min-height: 300px;
-    overflow: hidden;
-}
-
-.member {
-    width: 100%;
-    display: flex;
-    gap: 2rem;
-    align-items: center;
-    padding: 1.5rem;
-    position: absolute;
-    border-radius: 10px;
-    transition: 1s;
-}
-
-.member img {
-    width: 200px;
-    height: 200px;
-    object-fit: cover;
-}
-
-.member .member-details .name {
-    font-size: 2.5rem;
-    font-weight: 400;
-    color: #566bfc;
-}
-
-.member .member-details .role {
-    font-size: 1.5rem;
-    font-weight: 400;
-    margin-bottom: 1rem;
-}
-
-.member .member-details .description {
-    color: grey;
-    font-size: .9rem;
-    text-align: justify;
-}
-
-.dots {
-    display: flex;
-    justify-content: space-around;
-    height: 30px;
-    width: 200px;
-    align-items: center;
-    margin-top: 2rem;
-}
-
-.dots__dot {
-    all: unset;
-    height: 10px;
-    width: 30px;
-    border-radius: 20px;
-    background-color: rgb(223, 223, 223);
-    cursor: pointer;
-}
-
-.dots__dot--active {
-    background-color: #566bfc;
-}
-
-
-/* mailing */
-.mailing {
-    background-color: #566bfc;
-}
-
-.mailing .wrapper {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    padding: 2rem 0;
-    min-height: 400px;
-}
-
-.mailing h2 {
-    color: white;
-}
-
-.mailing h3 {
-    color: white;
-}
-
-.mailing-form {
-    background-color: white;
-    height: 3.5rem;
-    padding: .5rem;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    border-radius: 10px;
-    width: 90vw;
-    max-width: 450px;
-}
-
-.mailing-form input {
-    all: unset;
-    width: 100%;
-    height: 100%;
-}
-
-
-/* footer */
-
-.footer-top .wrapper {
-    display: flex;
-    padding: 2rem 2rem 7rem 2rem;
-    flex-wrap: wrap;
-    gap: 4rem;
-    /* justify-content: center; */
-    background-color: white;
-}
-
-.footer-top h2 {
-    font-size: 1.4rem;
-    font-weight: 500;
-    margin-bottom: .5rem;
-}
-
-.footer-top p,
-.footer-top a {
-    color: grey;
-    font-size: .9rem;
-    text-align: justify;
-}
-
-.f-t-about {
-    width: 300px;
-}
-
-.doubleMenu {
-    display: flex;
-    flex: 1;
-    gap: 5rem;
-    justify-content: space-evenly;
-}
-
-.footer-contact img {
-    width: 15px;
-}
-
-.footer-contact p {
-    display: flex;
-    align-items: center;
-    gap: .5rem;
-    margin-bottom: .6rem
-}
-
-.footer-bottom {
-    text-align: center;
-    padding: 1.5rem;
-    color: white;
-    background-color: #566bfc;
-}
-
-
-
-@media screen and (min-width: 1368px) {
-    .heroLeft h1 {
-        font-size: 10rem;
-        line-height: 8rem;
-    }
-}
