--- Login-/Website/tablets.css
@@ -1,68 +1,0 @@
-@media screen and (max-width:1050px) {
-    .topNav {
-        padding: 0 2rem;
-    }
-
-    .hamburger {
-        display: flex;
-    }
-
-    .mainMenu {
-        display: none;
-    }
-
-    .hero {
-        align-items: center;
-        position: relative;
-        overflow: hidden;
-    }
-
-    .heroLeft {
-        width: 100%;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        height: 100%;
-        justify-content: center;
-    }
-
-    .heroLeft h1 {
-        font-size: 17vw;
-        text-align: center;
-        line-height: 14vw;
-    }
-
-    .heroLeft .searchForm {
-        width: 100%;
-    }
-
-    .heroLeft .searchForm input {
-        width: 100%;
-    }
-
-    .heroRight {
-        display: none;
-    }
-
-    .circle {
-        position: absolute;
-        border-radius: 100%;
-    }
-
-    .circle:nth-child(1) {
-        height: 60vw;
-        width: 60vw;
-        bottom: -6vh;
-        left: -20vw;
-        background: linear-gradient(135deg, rgba(0, 0, 255, 0), #d8ddff);
-    }
-
-    .circle:nth-child(2) {
-        top: 10vh;
-        right: -10vw;
-        height: 45vw;
-        width: 45vw;
-        background: linear-gradient(45deg, rgb(0, 0, 211), #c9d0ff);
-    }
-
-}
